+++ src/Signal.lua
@@ -4,6 +4,14 @@
 	Handlers are fired in order, and (dis)connections are properly handled when
 	executing an event.
 ]]
+local __DEV__ = _G.__DEV__
+
+local _, FFlagRoduxRemoveConnectTraceback = xpcall(function()
+	return game:DefineFastFlag("RoduxRemoveConnectTraceback", false)
+end, function()
+	return true
+end)
+
 local function immutableAppend(list, ...)
 	local new = {}
 	local len = #list
@@ -38,7 +46,7 @@ Signal.__index = Signal
 function Signal.new(store)
 	local self = {
 		_listeners = {},
-		_store = store
+		_store = store,
 	}
 
 	setmetatable(self, Signal)
@@ -53,43 +61,48 @@ function Signal:connect(callback)
 
 	if self._store and self._store._isDispatching then
 		error(
-			'You may not call store.changed:connect() while the reducer is executing. ' ..
-				'If you would like to be notified after the store has been updated, subscribe from a ' ..
-				'component and invoke store:getState() in the callback to access the latest state. '
+			"You may not call store.changed:connect() while the reducer is executing. "
+				.. "If you would like to be notified after the store has been updated, subscribe from a "
+				.. "component and invoke store:getState() in the callback to access the latest state. "
 		)
 	end
 
 	local listener = {
 		callback = callback,
 		disconnected = false,
-		connectTraceback = debug.traceback(),
-		disconnectTraceback = nil
+		connectTraceback = nil,
+		disconnectTraceback = nil,
 	}
+
+	if not FFlagRoduxRemoveConnectTraceback or __DEV__ then
+		listener.connectTraceback = debug.traceback()
+	end
 
 	self._listeners = immutableAppend(self._listeners, listener)
 
 	local function disconnect()
 		if listener.disconnected then
-			error((
-				"Listener connected at: \n%s\n" ..
-				"was already disconnected at: \n%s\n"
-			):format(
-				tostring(listener.connectTraceback),
-				tostring(listener.disconnectTraceback)
-			))
+			error(
+				("Listener connected at: \n%s\n" .. "was already disconnected at: \n%s\n"):format(
+					tostring(listener.connectTraceback),
+					tostring(listener.disconnectTraceback)
+				)
+			)
 		end
 
 		if self._store and self._store._isDispatching then
 			error("You may not unsubscribe from a store listener while the reducer is executing.")
 		end
 
+		if not FFlagRoduxRemoveConnectTraceback or __DEV__ then
+			listener.disconnectTraceback = debug.traceback()
+		end
 		listener.disconnected = true
-		listener.disconnectTraceback = debug.traceback()
 		self._listeners = immutableRemoveValue(self._listeners, listener)
 	end
 
 	return {
-		disconnect = disconnect
+		disconnect = disconnect,
 	}
 end